+++ packages/select/src/index.js
@@ -1,6 +1,6 @@
-import { SelectProvider } from "./SelectProvider"
-import SelectInput from "./SelectInput"
-import SelectItem from "./SelectItem"
-import SelectList from "./SelectList"
+import { SelectProvider } from "./select-provider"
+import SelectInput from "./select-input"
+import SelectList from "./select-list"
+import SelectItem from "./select-item"
 
 export { SelectProvider as Select, SelectInput, SelectList, SelectItem }
